% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/precure.R
\docType{package}
\name{precure}
\alias{precure}
\alias{precure-package}
\title{Pretty Cure Color Palette for R}
\description{
Pretty Cure Color Palette for R
}
\details{
Palettes provide from Pretty Cure series that Japanese animation
}

