% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{precure_pal}
\alias{precure_pal}
\title{Precure all star's color palette}
\format{\preformatted{List of 12
 $ PretyCure      : chr [1:2] "#060825" "#FEFEFE"
 $ MaxHeart       : chr [1:3] "#1E1624" "#F5F2F6" "#F6F15B"
 $ SplashStar     : chr [1:2] "#EE3185" "#4F225D"
 $ YesPreCure5    : chr [1:5] "#E02C85" "#CA4241" "#1F5CAF" "#0F8355" ...
 $ YesPreCure5GoGo: chr [1:6] "#E62581" "#DE3228" "#F5E339" "#048161" ...
 $ Fresh          : chr [1:4] "#F9D1DE" "#357CC2" "#FDD549" "#B6203B"
 $ HeartCatch     : chr [1:4] "#F250B4" "#67D6EC" "#F9E35D" "#CEC5D8"
 $ Suite          : chr [1:4] "#D23B87" "#F4ED8C" "#7390ED" "#F0C32D"
 $ Smile          : chr [1:5] "#A61F68" "#E7350D" "#F9E04A" "#51CF51" ...
 $ DokiDoki       : chr [1:4] "#F578CB" "#595A83" "#F3A443" "#D7A5F3"
 $ HappinessCharge: chr [1:4] "#D02A8A" "#74A2EA" "#F4C542" "#9689EA"
 $ Princess       : chr [1:3] "#EA91C3" "#A3E5E7" "#E78717"
}}
\usage{
precure_pal
}
\description{
Precure all star's color palette
}
\keyword{datasets}

